Function to subset and convert phyloseq data

Function to subset and convert phyloseq data

Usage

convert_phylo(data, select)

Arguments

data

a phyloseq object

select

Name of OTU or taxa to select, must match taxa name in data

Value

A data.frame object, with elements W as the observed counts, M as the sequencing depth, and the sample data with their original names.

Examples

data(soil_phylo)
convert_phylo(soil_phylo, "OTU.43")
#>         W      M Plants DayAmdmt Amdmt ID Day
#> S009  350  89290      1       01     1  D   0
#> S204   74 194356      1       21     1  D   2
#> S112  300  66290      0       11     1  B   1
#> S247   70  35635      0       22     2  F   2
#> S026   43  49017      0       00     0  A   0
#> S023  215  36682      1       00     0  C   0
#> S201  105 171706      0       20     0  A   2
#> S133  167 153494      1       10     0  C   1
#> S103  164  73168      1       10     0  C   1
#> S012  117  31182      0       01     1  B   0
#> S134 1136 160612      1       11     1  D   1
#> S108  374  96114      1       10     0  C   1
#> S207  292  92273      0       21     1  B   2
#> S123  480 115443      1       10     0  C   1
#> S202  241 177721      0       21     1  B   2
#> S031  150  52771      0       00     0  A   0
#> S106  164 129553      0       10     0  A   1
#> S233   28  39284      1       20     0  C   2
#> S013  165  27713      1       00     0  C   0
#> S007   72  37877      0       01     1  B   0
#> S139  183 154182      1       11     1  D   1
#> S126   77 133446      0       10     0  A   1
#> S208  299 139063      1       20     0  C   2
#> S122  608 144272      0       11     1  B   1
#> S022  181  39898      0       01     1  B   0
#> S223  120  78864      1       20     0  C   2
#> S113  516  74789      1       10     0  C   1
#> S024   83  48814      1       01     1  D   0
#> S032  143  67299      0       01     1  B   0
#> S218   78  74627      1       20     0  C   2
#> S212  289 182268      0       21     1  B   2
#> S138  298  63448      1       10     0  C   1
#> S117   56  42691      0       11     1  B   1
#> S228  183 119662      1       20     0  C   2
#> S019  283  77203      1       01     1  D   0
#> S104   12 114784      1       11     1  D   1
#> S116  274  58551      0       10     0  A   1
#> S213  160  66819      1       20     0  C   2
#> S144  300 108677      0       12     2  F   1
#> S221  104  90967      0       20     0  A   2
#> S214  259  80154      1       21     1  D   2
#> S043   99  36098      0       02     2  F   0
#> S136  359  90128      0       10     0  A   1
#> S028  165  39352      1       00     0  C   0
#> S203  156 190471      1       20     0  C   2
#> S001  230  86007      0       00     0  A   0
#> S216   78  53720      0       20     0  A   2
#> S109   55  67615      1       11     1  D   1
#> S226  121 171724      0       20     0  A   2
#> S037   46  26554      0       01     1  B   0
#> S042   70  31996      0       02     2  F   0
#> S206   86  84638      0       20     0  A   2
#> S111  203  84124      0       10     0  A   1
#> S101  402 109686      0       10     0  A   1
#> S217  112  50293      0       21     1  B   2
#> S148   20  34617      0       12     2  F   1
#> S039   24  15038      1       01     1  D   0
#> S118  362 126421      1       10     0  C   1
#> S034  102  37348      1       01     1  D   0
#> S006   22  35309      0       00     0  A   0
#> S238  174  88192      1       20     0  C   2
#> S229   89  79837      1       21     1  D   2
#> S046  229  51708      0       02     2  F   0
#> S033   83  37283      1       00     0  C   0
#> S041   75  46249      0       02     2  F   0
#> S132  246 159371      0       11     1  B   1
#> S121   93  47684      0       10     0  A   1
#> S209   25  66337      1       21     1  D   2
#> S245   15  21660      0       22     2  F   2
#> S243   38  28671      0       22     2  F   2
#> S227   60  70308      0       21     1  B   2
#> S107  210  89310      0       11     1  B   1
#> S142  200 168244      0       12     2  F   1
#> S237   84  58807      0       21     1  B   2
#> S224   29  71940      1       21     1  D   2
#> S003  131  36728      1       00     0  C   0
#> S231  257  51393      0       20     0  A   2
#> S127   90 102935      0       11     1  B   1
#> S137  157  39955      0       11     1  B   1
#> S036   99  36097      0       00     0  A   0
#> S241   71  40139      0       22     2  F   2
#> S038   34   8830      1       00     0  C   0
#> S114  256  32298      1       11     1  D   1
#> S147   52  90944      0       12     2  F   1
#> S124  127  81038      1       11     1  D   1
#> S017  153  26555      0       01     1  B   0
#> S143  135 106468      0       12     2  F   1
#> S021   71  25060      0       00     0  A   0
#> S047   51  90848      0       02     2  F   0
#> S011   67  47710      0       00     0  A   0
#> S119  275 107738      1       11     1  D   1
#> S219   99  80119      1       21     1  D   2
#> S027   77  27587      0       01     1  B   0
#> S232   99  58540      0       21     1  B   2
#> S246   31  32120      0       22     2  F   2
#> S131  144  31203      0       10     0  A   1
#> S044   23  32861      0       02     2  F   0
#> S129   14  20512      1       11     1  D   1
#> S002   94  49719      0       01     1  B   0
#> S102  172  53477      0       11     1  B   1
#> S004   62  39569      1       01     1  D   0
#> S234  130  61928      1       21     1  D   2
#> S244  104  37578      0       22     2  F   2
#> S248   45  42810      0       22     2  F   2
#> S128   56  45576      1       10     0  C   1
#> S016   17  34523      0       00     0  A   0
#> S236  242  73434      0       20     0  A   2
#> S029  100  23329      1       01     1  D   0
#> S018   35  28479      1       00     0  C   0
#> S141   19  46890      0       12     2  F   1
#> S146   27  41880      0       12     2  F   1
#> S048   29  40087      0       02     2  F   0
#> S222   68  65700      0       21     1  B   2
#> S045   23  40986      0       02     2  F   0
#> S008  149  51542      1       00     0  C   0
#> S239   39  25681      1       21     1  D   2
#> S145   36  31203      0       12     2  F   1
#> S242    9  13813      0       22     2  F   2
#> S014  150  26968      1       01     1  D   0