Small soil phylum data for contrasts examples, otu table as data frame

A small subset of soil_phylo_otu used for examples of testing contrasts. A data frame made from the `phyloseq` object with only otu counts.

Usage

soil_phylum_contrasts_otu

Format

A phyloseq-class experiment-level object with an OTU table.

otu_table

OTU table with 39 taxa and 56 samples

References

Whitman, T., Pepe-Ranney, C., Enders, A., Koechli, C., Campbell, A., Buckley, D. H., Lehmann, J. (2016). Dynamics of microbial community composi-tion and soil organic carbon mineralization in soil following addition of pyrogenic andfresh organic matter. The ISME journal, 10(12):2918. <doi: 10.1038/ismej.2016.68>.